requirements.txt
# OpenAI Agents SDK Demo Requirements
openai>=1.12.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
requests>=2.31.0
pydantic>=2.5.0
aiohttp>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0